# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  response.ErrorData:
    properties:
      code:
        type: string
      details: {}
      error_trace:
        description: Full error + stack trace (non-prod only)
        type: string
      message:
        type: string
    type: object
  response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/response.ErrorData'
      success:
        example: false
        type: boolean
    type: object
...
  review.CreateReviewRequest:
    properties:
      comment:
        type: string
      context_id:
        type: integer
      context_type:
        type: string
      criteria:
        additionalProperties:
          type: integer
        type: object
      photos:
        items:
          type: string
        type: array
      rating:
        maximum: 5
        minimum: 1
        type: integer
      target_id:
        type: integer
      target_type:
        type: string
    required:
    - rating
    - target_id
    - target_type
    type: object
  review.OwnerResponseRequest:
    properties:
      response:
        type: string
    required:
    - response
    type: object
  review.Review:
    properties:
      attachments:
        description: Populated from attachment service (URLs)
        items:
          type: string
        type: array
      author_id:
        type: integer
      comment:
        type: string
      context_id:
        type: integer
      context_type:
        type: string
      created_at:
        type: string
      criteria:
        items:
          type: integer
        type: array
      id:
        type: integer
      is_hidden:
        type: boolean
      is_verified:
        type: boolean
      owner_response:
        type: string
      photos:
        description: 'DEPRECATED: use attachments WHERE target_type=''review_photos''
          AND target_id=id'
        items:
          type: string
        type: array
      rating:
        type: integer
      responded_at:
        type: string
      target_id:
        type: integer
      target_type:
        $ref: '#/definitions/review.TargetType'
      updated_at:
        type: string
    type: object
  review.TargetType:
    enum:
    - studio
    type: string
    x-enum-varnames:
    - TargetTypeStudio
  review.swaggerListReviewResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/review.Review'
        type: array
      success:
        type: boolean
    type: object
  review.swaggerReviewResponse:
    properties:
      data:
        $ref: '#/definitions/review.Review'
      success:
        type: boolean
    type: object
...
info:
  contact: {}
  description: API server for booking system.
  title: PhotoStudio API
  version: "1.0"
paths:
...
  /reviews:
    get:
      description: Возвращает список отзывов для определенной сущности (например,
        студии).
      parameters:
      - description: Тип сущности (например, studio)
        in: query
        name: target_type
        required: true
        type: string
      - description: ID сущности
        in: query
        name: target_id
        required: true
        type: integer
      - description: 'Лимит (дефолт: 20)'
        in: query
        name: limit
        type: integer
      - description: 'Отступ (дефолт: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список отзывов
          schema:
            $ref: '#/definitions/review.swaggerListReviewResponse'
        "400":
          description: Неверные параметры запроса
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Получить отзывы
      tags:
      - Reviews
    post:
      consumes:
      - application/json
      description: Пользователь может написать отзыв о сущности (студия, профиль).
        Для студий требуется завершенное бронирование.
      parameters:
      - description: Данные отзыва
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/review.CreateReviewRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Отыв создан
          schema:
            $ref: '#/definitions/review.swaggerReviewResponse'
        "400":
          description: Ошибка входных данных
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Оставление отзыва запрещено (нет брони и т.д.)
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Отзыв от этого пользователя уже существует
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Написать отзыв
      tags:
      - Reviews
  /reviews/{id}/response:
    post:
      consumes:
      - application/json
      description: Владелец сущности может оставить официальный ответ на отзыв.
      parameters:
      - description: ID отзыва
        in: path
        name: id
        required: true
        type: integer
      - description: Текст ответа
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/review.OwnerResponseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Ответ добавлен
          schema:
            $ref: '#/definitions/review.swaggerReviewResponse'
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Вложено запрещено оставлять ответ
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Отзыв не найден
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Ответить на отзыв
      tags:
      - Reviews
...
  /studios/{id}/reviews:
    get:
      description: Legacy эндпоинт для получения отзывов студии.
      parameters:
      - description: ID студии
        in: path
        name: id
        required: true
        type: integer
      - description: Лимит
        in: query
        name: limit
        type: integer
      - description: Отступ
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список отзывов
          schema:
            $ref: '#/definitions/review.swaggerListReviewResponse'
        "400":
          description: Некорректный ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Получить отзывы студии
      tags:
      - Reviews
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"